+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - changed: Replace text 'plugins' with 'providers' in Buy/Sell
 - changed: Tweak the boot background color on Android.
 - fixed: Incorrect wording when disabling a token
+- fixed: Min/max price label position on the Markets charts
 - fixed: Prevent setting invalid spending limits
 - fixed: Turning off 'Dark Mode' causes crash
 - removed: 'Add/Edit Tokens" option for token rows